requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
dj-rest-auth==6.0.0
Django==4.2.13
django-allauth==0.63.3
django-cors-headers==4.4.0
django-extensions==3.2.3
django-guardian==2.4.0
django-multiselectfield==0.1.13
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
guardian==0.2.2
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
mysqlclient==2.2.4
phonenumbers==8.13.40
pillow==10.4.0
pycountry==24.6.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
APScheduler==3.11.0
amqp==5.3.1
async-timeout==5.0.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
cssselect==1.2.0
cssutils==2.11.1
daphne==4.1.2
django-background-tasks==1.2.8
django-celery-beat==2.8.0
django-timezone-field==7.1
djangorestframework-simplejwt==5.3.1
gevent==24.11.1
greenlet==3.1.1
hyperlink==21.0.0
incremental==24.7.2
kombu==5.4.2
lxml==5.3.0
more-itertools==10.6.0
msgpack==1.1.0
premailer==3.10.0
prompt_toolkit==3.0.48
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.2.1
service-identity==24.2.0
six==1.17.0
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
tzlocal==5.2
vine==5.1.0
wcwidth==0.2.13
websocket==0.2.1
websockets==14.2
zope.event==5.0
zope.interface==7.2